# app/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Startup State Atlas API
  version: "1.0"
  summary: Utah's agent-readable map of startup companies and resources.
  description: |
    The Startup State Atlas API powers the Founder Navigator, the Utah
    Startup Map, and the agent-native CLI/MCP surfaces.

    ## Authentication

    Two authentication modes are supported:

    - **Better Auth session cookie** (`auth.session_token`) — set by
      `/sign-in` and `/sign-up`. Used by humans clicking through the
      web UI. Roles: `owner`, `goeo_admin`, `superadmin`.
    - **`X-Atlas-Admin-Token`** header — service-account-style token
      validated against `env.ATLAS_ADMIN_TOKEN`. Used by the local
      CLI, the local stdio MCP server, and any external machine
      client. Bypasses Better Auth and assumes a privileged caller.

    Read endpoints are open to both. Write endpoints accept either.
    For owner-edit on `PATCH /api/v1/companies/{slug}`, the session
    user id must match `companies.claimed_by_user_id`.

    ## Error envelope

    Every non-2xx response uses the same shape:

    ```json
    { "error": { "code": "string", "message": "string", "details": "unknown?" } }
    ```

    Codes used: `bad_request`, `not_found`, `forbidden`, `unauthorized`,
    `conflict`, `internal`.

    ## Wire format

    All wire fields are `snake_case`. Internal TS uses `camelCase` and
    converts at the boundary via `lib/api-codec.ts`.
servers:
  - url: https://startup.utah.gov
...
security:
  - cookieAuth: []
  - atlasAdminToken: []
  - {}
...
paths:
...
  /api/v1/intro-requests:
    get:
      tags: [intros]
      operationId: listIntroRequests
      summary: The caller's intro requests (outbound + inbound).
      description: |
        Session required. Returns:
        - `outbound`: all rows where the caller is the requester, any status.
        - `inbound`: rows targeted at the caller (their investor profile
          or claimed company), but only after admin moves them to
          `accepted` or `introduced`.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Outbound + inbound.
          content:
            application/json:
              schema:
                type: object
                required: [outbound, inbound]
                properties:
                  outbound:
                    type: array
                    items: { $ref: "#/components/schemas/IntroRequestOutbound" }
                  inbound:
                    type: array
                    items: { $ref: "#/components/schemas/IntroRequestInbound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [intros]
      operationId: createIntroRequest
      summary: Submit a new intro request to GOEO.
      description: |
        Body's `target` is a discriminated union — exactly one of
        `investor` or `company` may be specified. The requester's role
        is snapshotted at request time so the audit trail reflects what
        they were when they asked, not what they are now.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntroRequestCreate" }
      responses:
        "201":
          description: Pending.
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string, pattern: "^irq_[a-z0-9]+$" }
                  status: { type: string, enum: [pending] }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/intro-requests/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          pattern: "^irq_[a-z0-9]+$"
    get:
      tags: [intros]
      operationId: getIntroRequest
      summary: Read a single intro request.
      description: |
        Allowed for the requester, an admin, or the target party — but
        the target only sees the row after admin moves it to `accepted`
        or `introduced`.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The intro.
          content:
            application/json:
              schema:
                type: object
                required: [intro]
                properties:
                  intro: { $ref: "#/components/schemas/IntroRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [intros]
      operationId: reviewIntroRequest
      summary: Admin review — accept, decline, or mark introduced.
      description: |
        Admin only (`goeo_admin` / `superadmin`). Body's `status` must
        be one of `accepted | declined | introduced`.

        Concurrent-action guard: only `pending` rows are actionable.
        Two admins acting near-simultaneously will see one win and one
        get a 409.

        Email side-effects:
        - `accepted`: emails both parties with each other's contact info
          (or just the requester if the target hasn't claimed their
          profile, with a note that GOEO will follow up).
        - `declined`: emails the requester only.
        - `introduced`: emails the requester only (used when admin made
          the intro out-of-band).
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntroRequestPatch" }
      responses:
        "200":
          description: Updated intro.
          content:
            application/json:
              schema:
                type: object
                required: [intro]
                properties:
                  intro: { $ref: "#/components/schemas/IntroRequest" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Intro already moved past `pending`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
...
  responses:
    BadRequest:
      description: Validation error or malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: Authenticated but not authorized for this resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth.session_token
      description: |
        Better Auth session cookie. Set by `/sign-in` or `/sign-up`.
        Roles: `owner`, `goeo_admin`, `superadmin`. Owner edits gate
        on `companies.claimed_by_user_id` matching `user.id`.
    atlasAdminToken:
      type: apiKey
      in: header
      name: X-Atlas-Admin-Token
      description: |
        Service-account-style token validated against
        `env.ATLAS_ADMIN_TOKEN`. The local CLI (`npm run cli` from a
        checkout), the local stdio MCP server (`npm run mcp`), and
        external machine clients use this header for write operations.
        Bypasses Better Auth and assumes a privileged caller.
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: bad_request }
            message: { type: string }
            details: {}
...
    IntroRequestCreate:
      type: object
      required: [target, message_text]
      properties:
        target:
          oneOf:
            - type: object
              required: [type, id]
              properties:
                type: { type: string, enum: [investor] }
                id: { type: string, pattern: "^inv_[a-z0-9]+$" }
            - type: object
              required: [type, id]
              properties:
                type: { type: string, enum: [company] }
                id: { type: string, pattern: "^co_[a-z0-9]+$" }
        message_text:
          type: string
          minLength: 20
          maxLength: 2000
    IntroRequestPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, declined, introduced]
        admin_notes:
          type: string
          maxLength: 2000
          nullable: true
    IntroRequest:
      type: object
      required: [id, status, message_text, requester_role, target_type]
      properties:
        id: { type: string, pattern: "^irq_[a-z0-9]+$" }
        status:
          type: string
          enum: [pending, accepted, declined, introduced]
        message_text: { type: string }
        admin_notes: { type: string, nullable: true }
        submitted_at: { type: string, format: date-time, nullable: true }
        reviewed_at: { type: string, format: date-time, nullable: true }
        requester_role: { type: string }
        requester_user_id: { type: string }
        target_type:
          type: string
          enum: [investor, company]
        target_id: { type: string, nullable: true }
        target_name: { type: string, nullable: true }
        target_url: { type: string, nullable: true }
    IntroRequestOutbound:
      type: object
      required: [id, status]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [pending, accepted, declined, introduced]
        submitted_at: { type: string, format: date-time, nullable: true }
        reviewed_at: { type: string, format: date-time, nullable: true }
        admin_notes: { type: string, nullable: true }
        message_text: { type: string }
        target_type:
          type: string
          enum: [investor, company]
        target_id: { type: string, nullable: true }
        target_name: { type: string, nullable: true }
        target_url: { type: string, nullable: true }
    IntroRequestInbound:
      type: object
      required: [id, status]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [accepted, introduced]
        submitted_at: { type: string, format: date-time, nullable: true }
        reviewed_at: { type: string, format: date-time, nullable: true }
        admin_notes: { type: string, nullable: true }
        message_text: { type: string }
        requester_role: { type: string }
        requester_name: { type: string, nullable: true }
        requester_email: { type: string, format: email, nullable: true }